pyproject.toml
[project]
name = "shelf-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib==5.0.0",
    "arq==0.28.0",
    "bcrypt==4.3.0",
    "cashews[diskcache,redis]==7.4.4",
    "click==8.3.1",
    "fastapi==0.121.3",
    "filetype==1.2.0",
    "gunicorn==23.0.0",
    "httpx[http2]==0.28.1",
    "jinja2==3.1.6",
    "orjson==3.11.7",
    "passlib==1.7.4",
    "pi-heif==1.1.1",
    "pillow==12.0.0",
    "pydantic-settings==2.11.0",
    "pydantic[email]==2.12.5",
    "pymupdf==1.26.7",
    "python-jose[cryptography]==3.5.0",
    "python-multipart==0.0.22",
    "sentry-sdk==2.48.0",
    "stream-zip==0.0.84",
    "tortoise-orm[asyncpg]==1.1.7",
    "uvicorn[standard]==0.38.0",
    "uvloop==0.22.1",
]

[dependency-groups]
dev = [
    "ipython==9.12.0",
]
lint = [
    "mypy==1.20.1",
    "pre-commit==4.5.1",
]
test = [
    "aiosmtpd==1.4.6",
    "anyio==4.13.0",
    "faker==37.12.0",
    "freezegun==1.5.5",
    "pytest==8.4.2",
    "pytest-cov==7.1.0",
]
